Split `tests/bin/agent`
